# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "EmailSyncMode": {
        "enum": [
          "full_backfill",
          "incremental"
        ],
        "title": "EmailSyncMode",
        "type": "string"
      },
      "EmailSyncOptions": {
        "additionalProperties": false,
        "description": "User-selected bounds for a manual extraction run.",
        "properties": {
          "before_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Date"
          },
          "max_age_days": {
            "anyOf": [
              {
                "maximum": 3650.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Age Days"
          },
          "max_messages": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Messages"
          },
          "max_pages": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages"
          },
          "since_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since Date"
          }
        },
        "title": "EmailSyncOptions",
        "type": "object"
      },
...
      "EmailSyncStatus": {
        "description": "Current or last manual sync run status exposed at the API boundary.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "message_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Message Count",
            "type": "integer"
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailSyncMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "page_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Page Count",
            "type": "integer"
          },
          "progress": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailProviderName"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_email_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Raw Email Count",
            "type": "integer"
          },
          "recovered_from_expired_cursor": {
            "default": false,
            "title": "Recovered From Expired Cursor",
            "type": "boolean"
          },
          "retained_body_failure_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retained Body Failure Count",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/EmailSyncRunState"
          },
          "target_message_count": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Message Count"
          }
        },
        "required": [
          "state"
        ],
        "title": "EmailSyncStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Job Search Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sync": {
      "post": {
        "operationId": "sync_now_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/EmailSyncOptions"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Options"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSyncStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Sync Now",
        "tags": [
          "sync"
        ]
      }
    },
    "/sync/emails": {
      "get": {
        "description": "Returns a deterministic page of locally synced raw-email metadata within an optional sent-date window. Never contacts the email provider or triggers classification.",
        "operationId": "sync_emails_sync_emails_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sent_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sent After"
            }
          },
          {
            "in": "query",
            "name": "sent_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sent Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawEmailPreviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Synced Emails",
        "tags": [
          "sync"
        ]
      }
    },
    "/sync/emails/{public_id}/content": {
      "get": {
        "description": "Returns normalized plain-text content for one synced email. Retained bodies are served from local storage; metadata-only messages are fetched transiently from the email provider and are not persisted.",
        "operationId": "sync_email_content_sync_emails__public_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawEmailDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Read Synced Email Content",
        "tags": [
          "sync"
        ]
      }
    },
    "/sync/estimate": {
      "get": {
        "description": "Returns a deterministic local approximation of how much email a sync scope covers, using already-synced metadata only. It never calls the email provider.",
        "operationId": "sync_estimate_sync_estimate_get",
        "parameters": [
          {
            "in": "query",
            "name": "max_messages",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Messages"
            }
          },
          {
            "in": "query",
            "name": "since_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since Date"
            }
          },
          {
            "in": "query",
            "name": "before_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Date"
            }
          },
          {
            "in": "query",
            "name": "max_age_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 3650,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Age Days"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncScopeEstimate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Estimate Sync Scope",
        "tags": [
          "sync"
        ]
      }
    },
    "/sync/recent-emails": {
      "get": {
        "description": "Return recently stored raw-email metadata without body text.\n\n``order=sent_at`` (default) shows the newest synced mailbox messages;\n``order=ingested_at`` is the diagnostic view of what the latest sync wrote.",
        "operationId": "sync_recent_emails_sync_recent_emails_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RawEmailPreviewOrder",
              "default": "sent_at"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RawEmailPreviewRecord"
                  },
                  "title": "Response Sync Recent Emails Sync Recent Emails Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Recent Emails",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}